+++ src/main/java/com/example/iblog/services/ArticleService.java
@@ -2,12 +2,13 @@ package com.example.iblog.services;
 
 import com.example.iblog.domain.Article;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface ArticleService {
     public List<Article> getAll();
-    public Article getArticle(int articleId);
+    public Article getArticle(BigInteger articleId);
     public int insertArticle(Article article);
     public int updateArticle(Article article);
-    public int deleteArticleById(int articleId);
+    public int deleteArticleById(BigInteger articleId);
 }
